+++ CHANGELOG.md
@@ -4,8 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
-
----
+### Added
+- `Uncons`, for destructuring an `Iterable` into its head and tail
 
 ## [1.6.1] - 2017-06-17
 ### Changed